Revert "Add shared tool approval timeout constant (#1259)"
This reverts commit bcccfffa17314d533235cc6191670f2d50c1a2af.
