+++ packages/shared/src/lib/ai/tool-approval/index.ts
@@ -1,2 +1,7 @@
-export * from './constants';
-export * from './types';
+export type ToolStepContext = {
+  stepName: string;
+  actionDescription?: string;
+  actionName?: string;
+};
+
+export type ToolApprovalType = 'DefaultToolApproval' | 'RiskyStepToolApproval';
