--- packages/shared/src/lib/ai/tool-approval/types.ts
@@ -1,7 +1,0 @@
-export type ToolStepContext = {
-  stepName: string;
-  actionDescription?: string;
-  actionName?: string;
-};
-
-export type ToolApprovalType = 'DefaultToolApproval' | 'RiskyStepToolApproval';
